Atualiza método deleteById() em LivroService

+++ CRUD_02/livros/livros/src/main/java/app/service/LivroService.java
@@ -45,7 +45,11 @@ public class LivroService {
 	}
 
 	public String deleteById(Long id) {
-		this.livroRepository.deleteById(id);
-		return "Livro deletado com sucesso!";
+		if(this.findById(id) == null) {
+			return "Livro não encontrado!";
+		} else {
+		    this.livroRepository.deleteById(id);
+		    return "Livro deletado com sucesso!";
+		}
 	}
 }
